refactor: split signer success handlers
